Add files via upload
Selected Illustrations of Extinct Birds. Cropped from the original Files.
Source: "Extinct birds : an attempt to unite in one volume a short account of those birds which have become extinct in historical times : that is, within the last six or seven hundred years : to which are added a few which still exist, but are on the verge of extinction" By "Rothschild, Lionel Walter Rothschild, Baron, 1868-1937"
URL: https://www.biodiversitylibrary.org/bibliography/77436
Copyright & Usage according to BHL: "Copyright Status: Public domain. The BHL considers that this work is no longer under copyright protection."